rename sequenceidmanager file
